requirements.txt
gradio
langgraph
dspy-ai
plotly
networkx
requests
